requirements.txt
boto3==1.5.28
botocore==1.8.42
certifi==2018.1.18
chardet==3.0.4
confusable-homoglyphs==3.0.0
defusedxml==0.5.0
dj-database-url==0.4.2
Django==2.0.2
django-braces==1.12.0
django-cors-middleware==1.3.1
django-oauth-toolkit==1.0.0
django-registration==2.4.1
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.7.7
docutils==0.14
gunicorn==19.7.1
idna==2.6
jmespath==0.9.3
oauthlib==2.0.6
Pillow==5.0.0
psycopg2==2.7.3.2
PyJWT==1.5.3
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
s3transfer==0.1.12
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.6.0
urllib3==1.22
whitenoise==3.3.1
